move files and apply requested changes
