+++ src/components/explorer/ExplorerSearch.tsx
@@ -3,7 +3,7 @@ import { DropdownMenu } from 'components/common/dropdownMenu';
 import { ExplorerSearchDropdownButton } from 'components/explorer/ExplorerSearchDropdownButton';
 import { ExplorerSearchDropdownItems } from 'components/explorer/ExplorerSearchDropdownItems';
 import { ExplorerSearchInput } from 'components/explorer/ExplorerSearchInput';
-import ExplorerSearchSuggestions from 'components/explorer/ExplorerSearchSuggestions';
+import ExplorerSearchSuggestions from 'components/explorer/suggestion';
 import { ExplorerRouteGenerics } from 'components/explorer/utils';
 import { utils } from 'ethers';
 import { TradePair } from 'libs/modals/modals/ModalTradeTokenList';
