+++ src/components/explorer/ExplorerSearchInputContainer.tsx
@@ -5,20 +5,13 @@ import { ReactComponent as IconClose } from 'assets/icons/times.svg';
 
 type InputAttributes = JSX.IntrinsicElements['input'];
 interface Props
-  extends Pick<ExplorerSearchProps, 'setSearch' | 'search'>,
+  extends Pick<ExplorerSearchProps, 'search'>,
     Omit<InputAttributes, 'type'> {
   containerRef?: RefObject<HTMLDivElement>;
 }
 
 export const ExplorerSearchInputContainer: FC<Props> = (props) => {
-  const {
-    className,
-    containerRef,
-    children,
-    search,
-    setSearch,
-    ...inputProps
-  } = props;
+  const { className, containerRef, children, search, ...inputProps } = props;
   return (
     <div ref={containerRef} className="flex flex-grow">
       <input
@@ -28,7 +21,6 @@ export const ExplorerSearchInputContainer: FC<Props> = (props) => {
           'w-full flex-grow bg-transparent outline-none',
           className
         )}
-        onChange={(e) => setSearch(e.target.value)}
         {...inputProps}
       />
       {!!search && (
