+++ src/components/explorer/suggestion/index.ts
@@ -1,0 +1,4 @@
+import { memo } from 'react';
+import { SuggestionCombobox } from './SuggestionCombobox';
+
+export default memo(SuggestionCombobox);
